requirements-dev.txt
# Development-only packages
django-debug-toolbar==4.4.6
django-extensions==3.2.3
pydotplus>=2.0.2
graphviz>=0.21
